CI: Fix checkout on workflow
Due to the use of pull_request_target, the github_sha variable is always
the base branch one, and in the commit
469fac8a1accbf3a7090e9b0fc9206d50058dd75, we should check out to the PR
branch instead of the base branch.

This is because an external contributor can leak a write token, so for
that, I took the following steps:

- By default, the GIHUB_TOKEN is always read (As normal)
- On the coverage plugin, the token has access to write checks only.

So, in case of anyone want to leak the token, it's only read-only as any
normal PR. In the case of changing the actions themselves, it does not
matter because it is using the base branch.

Also, fix an issue with the format in the java code, to pass the test.

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>