Decouple examples from service
The purpose of this PR is to decouple the user's workflows from being
built into the workflow-service by using the PropertiesLauncher as the
Start-Class of the application.

Extra dependencies can be provided via system property `loader.path`.
In addition, to satisfy all runtime dependencies, the workflow-example
is being built with a plugin to include all of its dependencies in a
single jar.

For local development, a local profile was added to include the
workflow-examples dependency in the service's path:

```
mvn clean install -Plocal
```

Signed-off-by: Moti Asayag <masayag@redhat.com>