Enable graceful shutdown endpoint
The PR enables the shutdown endpoint which can be triggered by the
following when running with 'local' profile:

```
curl -u test:test  -X POST localhost:8080/actuator/shutdown
```

In addition, a script for local invocation of the workflow-service is
added. If the service dies, the script will relaunch it. It may be
useful for testing/development locally when there is a need to replace
only the workflows jar.

The graceful shutdown timeout
(spring.lifecycle.timeout-per-shutdown-phase) is set to 25s, since the
default timeout for K8s pods is 30s.
This can be revisit and tuned by changing the workflow-service
deployment to specify other timeout by terminationGracePeriodSeconds.

Signed-off-by: Moti Asayag <masayag@redhat.com>