Workflow: Properties annotation
This commit introduces a new way to annotate workflow metadata to the workflow
itself, the annotation comes with a single option, but with the idea to be
expanded in the near future

Each workflow now has a way to define a version that is stored in the
workflowDefinition.Properties column (JsonB) object. Users may want to determine the
version by hand, or using any tool, like the example given using
git-commit-id-plugin.

As an example:

```

@Bean(name = "simpleSequentialWorkFlow" + WorkFlowConstants.INFRASTRUCTURE_WORKFLOW)
@Infrastructure
@WorkFlowProperties(version = "${git.commit.id}")
WorkFlow simpleSequentialWorkFlowTask(@Qualifier("restCallTask") RestAPIWorkFlowTask restCallTask,
        @Qualifier("loggingTask") LoggingWorkFlowTask loggingTask) {
```

or:

```

@WorkFlowProperties(version = "v1.0.0")
```

The version is a string, but maybe we can think on semver here. I'm very open
here, but the idea was to use the commit id AFAIK, and not sure if it adds value
having in semver object.

The main reason to use Properties is that they can be extended with more fields
like: `@WorkFlowProperties(version = "v1.0.0", metricName="myAwesomeWorkflow")`
or some special logic like on_update, on_die, priority_key, labels.

Fix FLPATH-17

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>