Execute main workflow asynchronously
Since the duration of a workflow might be long, when a workflow is being
executed, it will be submitted to an executor and a response will return
the immediate "IN_PROGRESS" status with the execution-id to allow
monitoring of the progress by the caller.

Within the same patch, the async-executor of the continuation service is
being replaced in favor of a common executor for any async calls.

Signed-off-by: Moti Asayag <masayag@redhat.com>