prebuilt: Add tasks to analyze an application
Motivation
One of the steps to modernize an application incudes a proper analysis
that lists all the steps that must be taken. For example containerying
and application, that used to be deployed as a VM, means it no longer
can expect stuff to just exist on the file system, unless you mount the
right volumes, or some directories may not be missing entirely, or
special care for root access is needed.
The MTA[1] analysis is a tool that knows how to generate those reports
and that helps with the estimating the changes needed for the
application and planning that effort.

Modification:
Adding several tasks to create a workflow that can take a source code
repo (git url) and produce a reoprt with a containerization profile of
checks.
The tasks are:
  - CreateApplicationTask.java
  - GetApplicationTask.java
  - SubmitAnalysisTask.java
  - GetAnalysisTask.java

All tasks needs to get the URL of the MTA instance and the Bearer token
for authentication. Note, for testing puposes MTA can be deployed without
authentication.

Signed-off-by: Roy Golan <rgolan@redhat.com>