Add getWorkflowParameters API
The patch introduces a new API endpoint. The purpose of the new API is
to enable users to retrieve context parameters of an executed workflow,
specifically of type "assessment".

The patch adds a new controller method that maps to a new endpoint.
This new endpoint accepts the workflow execution ID and a list of context
parameters to retrieve. The method then retrieves the specified context
parameters from the workflow service and returns them as a JSON response.

An example of the request URL:
http://localhost:8080/api/v1/workflows/${workflowExecutionId}/context?param=WORKFLOW_OPTIONS

In the future, when additional parameters will be defined as 'public' visible, the URLcan be used as:
http://localhost:8080/api/v1/workflows/${workflowExecutionId}/context?param=WORKFLOW_OPTIONS,OTHER_PARAM
or a list of separate 'param's:
http://localhost:8080/api/v1/workflows/${workflowExecutionId}/context?param=WORKFLOW_OPTIONS&param=OTHER_PARAM

The response body includes the workflow execution ID and the retrieved
context parameters. An example of the response body:
{
  "workFlowExecutionId": "91811189-a70f-4bce-bd0b-6aa41e36986d",
  "workFlowOptions": {
    "newOptions": [
      {
        "identifier": "onboardingOption",
        "displayName": "Onboarding",
        "description": "An example of a complex WorkFlow",
        "details": [
          "An example workflow option of a complex workFlow with status checks"
        ],
        "workFlowName": "complexWorkFlow"
      }
    ]
  }
}

To retrieve context parameters, users will have to specify the parameters
they want to retrieve as defined in the WorkContextDelegate.Resource, for
resources that are defined as "public" visible. Currently, only the
WORKFLOW_OPTIONS context parameter is defined as "public" visible.

The schema change can be viewed in the generated openapi.json within the commit.

Signed-off-by: Moti Asayag <masayag@redhat.com>