Unify the setting/getting of project-id from context
WorkContext allows to set parameters on a map of type <string,Object>.
This map is being stored in table workflow_execution_context as JSON.
When the object is read, fields of type UUID returned as string.

To make sure we use same method to read and write the project ID from
the context, two methods are added.

Signed-off-by: Moti Asayag <masayag@redhat.com>