Use time.Instant instead of long
A long for comparing time is ok and very effiecient. However the way to
create one always involes using Date object.
The Instant object is immutable and have much expressive API for points
in time. Also, when using a long, you must name the variable with the
unit used, causing confusion and possibly bugs, and for sure longer var
names.

Signed-off-by: Roy Golan <rgolan@redhat.com>