MTA Assessment Workflow
This is an assessment workflow which is based on the Migration Analysis
infra workflow.
The main differne is that the checker of that workflow downloads the
html report and tries to extract a table of scores of issues found by
the analysis. The ProcessAnalysisTask has a predicate passed in the
constructor to test against the scores of the analysis and based on that
the task returns the passOption or the defaultOption. The passOption is
the next flow to invoke, that should be the move2kube.

The flow is roughly:

Start
  -
Infra flow start
  -
submit analysis report
  -
End

Start
  -
Checker flow start
  -
check report is ready
  -
switch (is ready?)
  yes -> go to parse
  no  -> go to check report is ready
  -
parse
  -
switch (match incidents on predicate)
  true - return pass option (move2kube)
  false - return default option
  -
End

Signed-off-by: Roy Golan <rgolan@redhat.com>