A notifier bean for interfacing with the notification service
Tasks of workflows need a builtin conveient way of sending notifications
and for a lot of reasons it makes more sense then a special notificaion
tasks - specially when the input for the message exists in the task
context and its waste or just too much to pass on the context for a
special notification task. It also doesn't need plumbing the task into a
workflow which is combersome.

A NotificationTask probably still makes sense for user-defined
notifications.

To send a notificaion just inject a Notifier into a bean:

@Bean
SomeTask createSomeTask(Notifier notifier) {
      new SomeTask(notifier)...
}

class SomeTask {
    SomeTask(Notifier notifier) { this.notifier = notifier; ... }

    execute(WorkContext ctx) {
        ... //work
        var m = new NotificationMessageCreateRequestDTO();
        m.setBody("body");
        notifier.send(m);
    }
}

Signed-off-by: Roy Golan <rgolan@redhat.com>