Add delay between attempts of invoking requests
With the current implementation, the client intensively sends requests
to the workflow service if the former request failed.
This intensity leads to high CPU consumption which impacts both client
and mainly the server.

By adding a small timeout between failed requests, the chance for
completing a workflow increases and so the resource consumption drops
significantly (20% CPU vs over 100%).

This is a temporary fix before replacing the current implementation with
Executor service and future.

Signed-off-by: Moti Asayag <masayag@redhat.com>