Propagte security context to threads created by spring
Motivation
All flow and task execution are async (@Async) and All
of them are lacking the security context. This means sending
notification or any other action that needs this info is lacking it.

Modification
Configure the context strategy on bootup time so every thread that
spring creates will inherit the security context.

Result
Sending notifications using the Notifier interface doesn't need any user
information

Gaps
Workflows that are continued upon server restart are probably missing
this information and we need to probably fetch the user info from the DB
and rebuild the security context.

Signed-off-by: Roy Golan <rgolan@redhat.com>