Match SDK generators
There is a different behavior between the two SDK generators,
e.g. the notification service will always try to generate the SDK,
even if it isn't needed.

This PR aligns the output of the two SDK generators to look and
behave similarly. However, a few changes remain (e.g. the path
to the generated openapi.json file and also the package names
which will be updated in next PR).

Signed-off-by: Moti Asayag <masayag@redhat.com>