Add invokingExecutionId to WorkflowRequestDTO to support context passing
Enable using arguments from different another execution to use its
execution level parameters.

This could be used when an assessment workflow gathers inormation and
stores it on the context and then when invoking a follow up workflow
from the options of that assessment then the requester can pass the
execution ID of the assessment.

The data flow looks like that:

Start -> Asssessment flow -> addParameter "foo": "bar" to WorkContext -> End with
option 1

Start -> Workflow from option 1 -> use Workcontext param "foo" -> work
-> End

The invoking execution ID isn't stored in the DB, it is just a request
argument passed in the POST request for invoking a workflow.

Signed-off-by: Roy Golan <rgolan@redhat.com>