`prebuilt-tasks` module must NOT depend on `workflow-service-sdk`
This module (prebuilt-tasks) must not depend on `workflow-service-sdk`,
as it is already dependent on `workflow-service`. `workflow-service`
depends on `prebuilt-tasks`. Adding `workflow-service-sdk` here will
create a circular dependency, and `workflow-service` won't be able to
generate the `openapi.json` file anymore.

This commit introduces a check that will now prevent the build from succeeding if there is a dependency on workflow-service-sdk.

Signed-off-by: Gloria Ciavarrini <gciavarrini@redhat.com>